regression bugs in map reverted

+++ map.c
@@ -1927,7 +1927,7 @@ static int16_t map_test_get_ratio(void) {
 	delete();
 	printf("map deleted\n");
 	int16_t all_ok = 1;
-	uint16_t result = map_get_ratio(-100, -100, 100, 100, 100, 0, 0);
+	uint8_t result = map_get_ratio(-100, -100, 100, 100, 100, 0, 0);
 	if (result != 255) {
 		all_ok = 0;
 	}
@@ -1967,7 +1967,7 @@ static int16_t map_test_get_ratio(void) {
 		int8_t ok = 1;
 		for (i=-150; i<=150; i+=50) {
 			for (j=-150; j<=150; j+=50) {
-				int16_t last_result = map_get_ratio(i, j, 155, 155, 0, value, value);
+				uint8_t last_result = map_get_ratio(i, j, 155, 155, 0, value, value);
 				if (last_result != 255) {
 					printf("\tmap_get_ratio(%d, %d, 155, 155, %u, %d, %d) = %u\n", i, j, width, value, value, last_result);
 					printf("test(%4d,%4d)\tFAILED\n", i, j);
@@ -1975,7 +1975,7 @@ static int16_t map_test_get_ratio(void) {
 				}
 				for (width=0; width<=6400; width+=8) {
 					result = map_get_ratio(i, j, 155, 155, width, value, value);
-					int16_t diff = abs((int)result - last_result);
+					int16_t diff = abs((int16_t)result - (int16_t)last_result);
 					if (diff > 2) {
 						ok = 0;
 						printf("\tmap_get_ratio(%d, %d, 155, 155, %u, %d, %d) = %u\n", i, j, width, value, value, result);
